update some translations, introduce s3 iam authentication and dialogs boilerplate
