+++ package.json
@@ -82,6 +82,7 @@
     "@sentry/browser": "^5.15.4",
     "any-shell-escape": "^0.1.1",
     "aws-amplify": "^3.0.6",
+    "aws-sign": "^1.0.0",
     "axios": "^0.19.2",
     "bent": "^7.1.0",
     "brace": "^0.11.1",
