+++ src/components/ImportFromCognitoS3Dialog/set-type-of-file-to-load-and-disable.js
