+++ src/i18n/locales/cn/translations.json
@@ -26,7 +26,7 @@
   "sample-brushes": "刷子样品",
   "complete": "完整",
   "review": "回顾",
-  "collobrate": "协作",
+  "collaborate": "协作",
   "join-a-session": "加入会议",
   "leave-session": "休会",
   "next": "下一个",
