+++ src/i18n/locales/en/translation.json
@@ -28,7 +28,7 @@
   "sample-brushes": "Sample Brushes",
   "complete": "Complete",
   "review": "Review",
-  "collobrate": "Collaborate",
+  "collaborate": "Collaborate",
   "join-a-session": "Join a Session",
   "leave-session": "Leave Session",
   "next": "Next",
@@ -57,6 +57,9 @@
   "import-text-snippets": "Import Text Snippets",
   "import-toy-dataset": "Import Toy Dataset",
   "import-from": "Import from",
+  "import-from-cognito-s3": "Import from S3 (Cognito)",
+  "import-from-s3": "Import from S3 (IAM)",
+  "upload-to-s3": "Upload via S3",
   "name": "Name",
   "type": "Type",
   "size": "Size",
