+++ src/i18n/locales/fr/translations.json
@@ -26,7 +26,7 @@
   "sample-brushes": "Brosses à échantillons",
   "complete": "Compléter",
   "review": "Revue",
-  "collobrate": "Collaborer",
+  "collaborate": "Collaborer",
   "join-a-session": "Participer à une session",
   "leave-session": "Séance de congé",
   "next": "Suivant",
